Fix handling of profile-name
The current code:
self.profile_name = op.get('crush-profile', profile_name)

will only default to profile_name if the 'crush-profile' key
doesn't exist in the op dictionary. If the 'crush-profile' key
exists and is set to None, the default profile_name is not used.

This change will use the default profile_name in both cases.

A full charm-helpers sync is done here.

Closes-Bug: #1960622
Change-Id: If9749e16eadfab5523d06c82f3899a83b8c6fdc1