Rewrite actions/copy_pool into the oeprator framework
In addition to trivial changes (passing `event` into
the `copy_pool` function), this change introduces an
update to the actions/__init__.py that allows succinct
import and use from the main charm.py.

An apparently unrelated change is the removal of
charm-proof from the lint job, as it fails with the
removal of actions/copy-pool.

Change-Id: I66a5590ddf0f0bb5ca073a91b451f8c78598609a
func-test-pr: https://github.com/openstack-charmers/zaza-openstack-tests/pull/866