Fix issue with ceph-client relation handling
A bug was introduced when changing ceph-client to
an operator framework library that caused the
fallback application_name handling to present
a class name rather than a remote applicaiton name.

This change updates the handling to get at an
`app.name` rather than an `app`.

As a drive-by, this also allow-lists the fully-
qualified rename.sh.

Closes-Bug: #1995086
Change-Id: I57b685cb78ba5c4930eb0fa73d7ef09d39d73743
func-test-pr: https://github.com/openstack-charmers/zaza-openstack-tests/pull/1022